(Hedgehog) Lost World Path Spline Fix
My OBJ imports crashed if I didn't have write the path offset table thing, so I've just flipped around the check for it.

Also made the OBJ import looks for a `Scale Modifier = ` string in the same line it checks for the identifier in, as a lot of tool stuff depend on the Generations units, which Lost World and Forces don't use.