Use new CheckValue function
Rather than just skipping a number I believe is always the same, this (if in Debug) checks it (which might be like a millisecond slower or something but who cares?) and throws an exception if it's wrong.

Haven't properly tested this, so I'll need to put myself together some sample files to check these edits don't fall flat on their face.